+++ src/main/kotlin/hilfling/backend/hilfling/controller/SecurityLevelController.kt
@@ -10,7 +10,7 @@ class SecurityLevelController {
     val repository = SecurityLevelRepository()
 
     @GetMapping("/{id}")
-    fun getSecurityLevelById(@PathVariable("id") id: Long): SecurityLevel? {
+    fun getSecurityLevelById(@PathVariable("id") id: Int): SecurityLevel? {
         return repository.findById(id)
     }
